+++ plugin/src/test/java/io/jenkins/plugins/casc/vault/VaultTestUtil.java
@@ -26,10 +26,11 @@ class VaultTestUtil {
     public static final String VAULT_PATH_KV2_1 = "kv-v2/admin";
     public static final String VAULT_PATH_KV2_2 = "kv-v2/dev";
     public static final String VAULT_PATH_KV2_3 = "kv-v2/qa";
+    public static final String VAULT_PATH_KV2_AUTH_TEST = "kv-v2/auth-test";
     public static String VAULT_APPROLE_ID = "";
     public static String VAULT_APPROLE_SECRET = "";
 
-    private static void runCommand(VaultContainer container, final String... command)
+    public static void runCommand(VaultContainer container, final String... command)
         throws IOException, InterruptedException {
         LOGGER.log(Level.FINE, String.join(" ", command));
         container.execInContainer(command);
@@ -73,7 +74,7 @@ class VaultTestUtil {
             // Create AppRole
             runCommand(container, "vault", "auth", "enable", "approle");
             runCommand(container, "vault", "write", "auth/approle/role/admin",
-                    "secret_id_ttl=10m", "token_num_uses=0", "token_ttl=20m", "token_max_ttl=20m",
+                    "secret_id_ttl=10m", "token_num_uses=0", "token_ttl=4s", "token_max_ttl=4s",
                     "secret_id_num_uses=1000", "policies=admin");
 
             // Retrieve AppRole credentials
@@ -91,6 +92,7 @@ class VaultTestUtil {
             runCommand(container, "vault", "kv", "put", VAULT_PATH_KV2_1, "key1=123", "key2=456");
             runCommand(container, "vault", "kv", "put", VAULT_PATH_KV2_2, "key3=789");
             runCommand(container, "vault", "kv", "put", VAULT_PATH_KV2_3, "key2=321");
+            runCommand(container, "vault", "kv", "put", VAULT_PATH_KV2_AUTH_TEST, "key1=auth-test");
 
         } catch (Exception e) {
             LOGGER.log(Level.WARNING, e.getMessage());